+++ src/MicroElements.Metadata/Metadata/IPropertyCalculator.cs
@@ -1,4 +1,7 @@
-﻿namespace MicroElements.Metadata
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace MicroElements.Metadata
 {
     /// <summary>
     /// Property value calculator.
@@ -10,7 +13,8 @@
         /// Calculates value.
         /// </summary>
         /// <param name="propertyContainer">Source property container.</param>
+        /// <param name="searchOptions">SearchOptions provided by client.</param>
         /// <returns>Value and <see cref="ValueSource"/>.</returns>
-        (T Value, ValueSource ValueSource) Calculate(IPropertyContainer propertyContainer);
+        (T Value, ValueSource ValueSource) Calculate(IPropertyContainer propertyContainer, SearchOptions searchOptions);
     }
 }