+++ src/MicroElements.Metadata/Metadata/PropertyCalculator.cs
@@ -12,14 +12,24 @@ namespace MicroElements.Metadata
     /// <typeparam name="T">Value type.</typeparam>
     public class PropertyCalculator<T> : IPropertyCalculator<T>
     {
-        private readonly Func<IPropertyContainer, T> _calculateSimple;
-        private readonly Func<IPropertyContainer, (T Value, ValueSource ValueSource)> _calculate;
+        private readonly Func<IPropertyContainer, SearchOptions, T> _calculateSimple;
+        private readonly Func<IPropertyContainer, SearchOptions, (T Value, ValueSource ValueSource)> _calculate;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PropertyCalculator{T}"/> class.
         /// </summary>
         /// <param name="calculateSimple">Calculate func.</param>
         public PropertyCalculator(Func<IPropertyContainer, T> calculateSimple)
+        {
+            calculateSimple.AssertArgumentNotNull(nameof(calculateSimple));
+            _calculateSimple = (container, options) => calculateSimple(container);
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PropertyCalculator{T}"/> class.
+        /// </summary>
+        /// <param name="calculateSimple">Calculate func.</param>
+        public PropertyCalculator(Func<IPropertyContainer, SearchOptions, T> calculateSimple)
         {
             _calculateSimple = calculateSimple.AssertArgumentNotNull(nameof(calculateSimple));
         }
@@ -28,21 +38,21 @@ namespace MicroElements.Metadata
         /// Initializes a new instance of the <see cref="PropertyCalculator{T}"/> class.
         /// </summary>
         /// <param name="calculate">Calculate func.</param>
-        public PropertyCalculator(Func<IPropertyContainer, (T Value, ValueSource ValueSource)> calculate)
+        public PropertyCalculator(Func<IPropertyContainer, SearchOptions, (T Value, ValueSource ValueSource)> calculate)
         {
             _calculate = calculate.AssertArgumentNotNull(nameof(calculate));
         }
 
         /// <inheritdoc />
-        public (T Value, ValueSource ValueSource) Calculate(IPropertyContainer propertyContainer)
+        public (T Value, ValueSource ValueSource) Calculate(IPropertyContainer propertyContainer, SearchOptions searchOptions)
         {
             if (_calculateSimple != null)
             {
-                T value = _calculateSimple(propertyContainer);
+                T value = _calculateSimple(propertyContainer, searchOptions);
                 return (value, ValueSource.Calculated);
             }
 
-            var calculationResult = _calculate(propertyContainer);
+            var calculationResult = _calculate(propertyContainer, searchOptions);
             return calculationResult;
         }
     }